requirements.txt
yt_dlp>=2024.4.9
openai-whisper>=20231117
tqdm>=4.66.1
numpy<2.0
ffmpeg-python>=0.2.0
torch>=2.0.0
torchaudio>=2.0.1
Pillow>=11.0.0
tk
